fix: address PR review feedback
- Fix skippedCount to also count updated deps with missing base analysis
- Fix npm token precedence: CLI > NPM_TOKEN env > config (was skipping env)
- Escape regex metacharacters in exclude patterns to prevent unexpected matches
- Extract exclude logic into shared src/utils/exclude.ts
- Wire excludePackages and npmToken through analyze command (was missing)
- Quote exclude-packages in action.yml to prevent bash glob expansion
- Add skippedCount: 0 to reporter.ts DiffSummary constructor
