+++ action.yml
@@ -68,7 +68,7 @@ runs:
           ARGS="$ARGS --security"
         fi
         if [ -n "${{ inputs.exclude-packages }}" ]; then
-          ARGS="$ARGS --exclude-packages ${{ inputs.exclude-packages }}"
+          ARGS="$ARGS --exclude-packages '${{ inputs.exclude-packages }}'"
         fi
 
         if [ -n "$VERSION" ]; then
