+++ src/cli/commands/analyze.ts
@@ -17,6 +17,7 @@ import { scanPhantomDependencies } from "../../core/phantom/scanner.js";
 import type { PinningReport } from "../../core/pinning/analyzer.js";
 import { analyzeVersionPinning } from "../../core/pinning/analyzer.js";
 import { setNpmCacheEnabled, setNpmToken } from "../../core/registry/npm-client.js";
+import { parseExcludePackages } from "../../utils/exclude.js";
 import { buildReport } from "../../core/report/builder.js";
 import type { Report } from "../../core/report/types.js";
 import { getDatabase } from "../../core/store/database.js";
@@ -65,8 +66,10 @@ export async function analyzeCommand(target: string, options: AnalyzeOptions): P
     setNpmCacheEnabled(false);
   }
 
-  if (options.npmToken) {
-    setNpmToken(options.npmToken);
+  // Set npm token: CLI option > env var > config
+  const resolvedNpmToken = options.npmToken ?? process.env.NPM_TOKEN ?? config.npmToken;
+  if (resolvedNpmToken) {
+    setNpmToken(resolvedNpmToken);
   }
 
   const isCi = options.ci || false;
@@ -199,11 +202,13 @@ async function analyzeFile(
   }
 
   // Use buildReportFromPackageJson for file analysis
+  const excludePackages = parseExcludePackages(options.excludePackages, config.excludePackages);
   const result = await buildReportFromPackageJson(pkg, {
     depth: options.depth,
     concurrency: options.concurrency,
     dev: options.dev,
     noCache: options.noCache,
+    excludePackages,
   });
 
   if (spinner) {
